+++ packages/core/src/wrap-in-schema.ts
@@ -1,5 +1,6 @@
 import React, {forwardRef, useContext, useState} from 'react';
 import type {SequenceControls} from './CompositionManager.js';
+import {flattenActiveSchema, type ResolveValue} from './flatten-schema.js';
 import type {
 	SchemaKeysRecord,
 	SequenceSchema,
@@ -22,30 +23,6 @@ const getNestedValue = (obj: Record<string, unknown>, key: string): unknown => {
 	}
 
 	return current;
-};
-
-type ResolveValue = (key: string) => unknown;
-
-const flattenActiveSchema = (
-	schema: SequenceSchema,
-	resolve: ResolveValue,
-): SequenceSchema => {
-	const out: SequenceSchema = {};
-	for (const key of Object.keys(schema)) {
-		const field = schema[key];
-		if (field.type === 'enum') {
-			out[key] = field;
-			const current = (resolve(key) as string | undefined) ?? field.default;
-			const variant = field.variants[current];
-			if (variant) {
-				Object.assign(out, flattenActiveSchema(variant, resolve));
-			}
-		} else {
-			out[key] = field;
-		}
-	}
-
-	return out;
 };
 
 const mergeValues = (
