+++ packages/studio-shared/src/index.ts
@@ -95,7 +95,12 @@ export {
 	UiOpenGlOptions,
 } from './render-job';
 export type {CompletedClientRender} from './render-job';
-export type {SchemaFieldInfo} from './schema-field-info';
+export {
+	SCHEMA_FIELD_ROW_HEIGHT,
+	UNSUPPORTED_FIELD_ROW_HEIGHT,
+	getSchemaFields,
+} from './schema-field-info';
+export type {SchemaFieldInfo, SequenceControls} from './schema-field-info';
 export {SOURCE_MAP_ENDPOINT} from './source-map-endpoint';
 export {
 	ScriptLine,
